Implement ttyname using /dev/fd links
