+++ tests/test_vlibc.c
@@ -640,6 +640,21 @@ static const char *test_ttyname_dev_tty(void)
     return 0;
 }
 
+static const char *test_ttyname_openpty(void)
+{
+    int m, s;
+    char expect[64];
+    mu_assert("openpty", openpty(&m, &s, expect, NULL, NULL) == 0);
+    char buf[64];
+    int r = ttyname_r(s, buf, sizeof(buf));
+    char *name = ttyname(s);
+    close(m);
+    close(s);
+    mu_assert("ttyname_r openpty", r == 0 && strcmp(buf, expect) == 0);
+    mu_assert("ttyname openpty", name && strcmp(name, expect) == 0);
+    return 0;
+}
+
 static const char *test_udp_send_recv(void)
 {
     int s1 = socket(AF_INET, SOCK_DGRAM, 0);
@@ -4644,6 +4659,7 @@ static const char *all_tests(void)
     mu_run_test(test_byte_order);
     mu_run_test(test_isatty_stdin);
     mu_run_test(test_ttyname_dev_tty);
+    mu_run_test(test_ttyname_openpty);
     mu_run_test(test_socket);
     mu_run_test(test_socketpair_basic);
     mu_run_test(test_socket_addresses);
